+++ Calliterra/src/Platform/DX11/DX11VertexBuffer.h
@@ -6,8 +6,8 @@
 class DX11VertexBuffer : public VertexBuffer
 {
 public:
-	DX11VertexBuffer(DX11Context& context, float* vertices, size_t elementCount);
-	DX11VertexBuffer(DX11Context& context, float** listOfVertexArrays, size_t* listOfElementCounts, size_t bufferCount);
+	DX11VertexBuffer(DX11Context& context, float* vertices, uint32_t elementCount);
+	DX11VertexBuffer(DX11Context& context, float** listOfVertexArrays, uint32_t* listOfElementCounts, uint32_t bufferCount);
 
 	void Bind() override;
 	void Unbind() override;
@@ -24,8 +24,8 @@ private:
 	//std::vector<ComPtr<ID3D11InputLayout>> m_D3DBufferLayoutList;
 	//std::vector<VertexBufferLayout> m_BufferLayoutList;
 	float** m_VertexArrayList;
-	size_t* m_ElementCountList;
-	size_t  m_BufferCount;
+	uint32_t * m_ElementCountList;
+	uint32_t m_BufferCount;
 
 	ComPtr<ID3D11Buffer> m_D3DVertexBufferArray[D3D11_IA_VERTEX_INPUT_RESOURCE_SLOT_COUNT] = {};
 	ComPtr<ID3D11InputLayout> m_D3DBufferLayoutArray[D3D11_IA_VERTEX_INPUT_RESOURCE_SLOT_COUNT] = {};
